AIP 4110: How to retrieve tokens from the metadata server
Add a complete discussion of how auth libraries and workloads should
cache and refresh tokens when loading them from the metadata server.

Existing GCP auth libraries handle this properly, but occasionally it is
not possible or desirable to use a preexisting auth library.  This new
section of the AIP provides guidance for those who need to directly
retrieve tokens.